IndexedDB: Chromium coding-style cleanup of new back-end c/o cpplint
BUG=234278
R=alecflett@chromium.org,dgrogan@chromium.org

Review URL: https://chromiumcodereview.appspot.com/16005006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@204065 0039d316-1c4b-4281-b951-d872f2087c98

+++ content/browser/indexed_db/indexed_db_cursor.h
@@ -6,6 +6,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #ifndef CONTENT_BROWSER_INDEXED_DB_INDEXED_DB_CURSOR_H_
 #define CONTENT_BROWSER_INDEXED_DB_INDEXED_DB_CURSOR_H_
 
+#include "base/basictypes.h"
 #include "base/memory/ref_counted.h"
 #include "content/common/indexed_db/indexed_db_key.h"
 
@@ -15,7 +16,7 @@ class IndexedDBCallbacksWrapper;
 
 class IndexedDBCursor : public base::RefCounted<IndexedDBCursor> {
  public:
-  virtual void Advance(unsigned long count,
+  virtual void Advance(uint32 count,
                        scoped_refptr<IndexedDBCallbacksWrapper> callbacks) = 0;
   virtual void ContinueFunction(
       scoped_ptr<IndexedDBKey> key,
